requirements.txt
flask==2.2.5
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Dance==7.0.0
Flask-Migrate==4.0.5
Flask-Minify==0.42
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
flask_wtf==1.2.1
azure-storage-blob==12.17.0
azure-functions==1.17.0
